+++ src/main/java/io/milvus/client/DescribeIndexResponse.java
@@ -29,7 +29,7 @@ public class DescribeIndexResponse {
   private final Response response;
   private final Index index;
 
-  public DescribeIndexResponse(Response response, @Nullable Index index) {
+  DescribeIndexResponse(Response response, @Nullable Index index) {
     this.response = response;
     this.index = index;
   }
@@ -45,6 +45,10 @@ public class DescribeIndexResponse {
 
   public Response getResponse() {
     return response;
+  }
+
+  public boolean ok() {
+    return response.ok();
   }
 
   @Override