+++ src/main/java/io/milvus/client/DescribeTableResponse.java
@@ -30,7 +30,7 @@ public class DescribeTableResponse {
   private final Response response;
   private final TableSchema tableSchema;
 
-  public DescribeTableResponse(Response response, @Nullable TableSchema tableSchema) {
+  DescribeTableResponse(Response response, @Nullable TableSchema tableSchema) {
     this.response = response;
     this.tableSchema = tableSchema;
   }
@@ -46,6 +46,10 @@ public class DescribeTableResponse {
 
   public Response getResponse() {
     return response;
+  }
+
+  public boolean ok() {
+    return response.ok();
   }
 
   @Override