+++ src/main/java/io/milvus/client/InsertResponse.java
@@ -28,7 +28,7 @@ public class InsertResponse {
   private final Response response;
   private final List<Long> vectorIds;
 
-  public InsertResponse(Response response, List<Long> vectorIds) {
+  InsertResponse(Response response, List<Long> vectorIds) {
     this.response = response;
     this.vectorIds = vectorIds;
   }
@@ -39,6 +39,10 @@ public class InsertResponse {
 
   public Response getResponse() {
     return response;
+  }
+
+  public boolean ok() {
+    return response.ok();
   }
 
   @Override