+++ src/main/java/io/milvus/client/MilvusClient.java
@@ -22,7 +22,7 @@ package io.milvus.client;
 /** The Milvus Client Interface */
 public interface MilvusClient {
 
-  String clientVersion = "0.2.2";
+  String clientVersion = "0.3.0";
 
   /** @return the current Milvus client version */
   default String getClientVersion() {
@@ -39,7 +39,11 @@ public interface MilvusClient {
    * ConnectParam connectParam = new ConnectParam.Builder()
    *                                             .withHost("localhost")
    *                                             .withPort("19530")
-   *                                             .withTimeout(10000)
+   *                                             .withConnectTimeout(10, TimeUnit.SECONDS)
+   *                                             .withKeepAliveTime(Long.MAX_VALUE, TimeUnit.NANOSECONDS)
+   *                                             .withKeepAliveTimeout(20, TimeUnit.SECONDS)
+   *                                             .keepAliveWithoutCalls(false)
+   *                                             .withIdleTimeout(24, TimeUnit.HOURS)
    *                                             .build();
    * </code>
    * </pre>
@@ -52,7 +56,8 @@ public interface MilvusClient {
    */
   Response connect(ConnectParam connectParam) throws ConnectFailedException;
 
-  /** @return <code>true</code> if the client is connected to Milvus server */
+  /** @return <code>true</code> if the client is connected to Milvus server.
+   * The channel's connectivity state is READY.*/
   boolean isConnected();
 
   /**